# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
  models.ActivateRequest:
    properties:
      authorization_code:
        description: 授权码，必填
        type: string
      device_info:
        additionalProperties: true
        description: 设备信息，可选
        type: object
      hardware_fingerprint:
        description: 硬件指纹，必填
        type: string
      software_version:
        description: 软件版本，可选
        type: string
    required:
    - authorization_code
    - hardware_fingerprint
    type: object
  models.ActivateResponse:
    properties:
      heartbeat_interval:
        description: 心跳间隔(秒)
        type: integer
      license_file:
        description: base64编码的加密许可证文件
        type: string
      license_key:
        description: 许可证密钥
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: 业务错误码，如 "100001", "900001"
        type: string
      message:
        description: 错误描述信息
        type: string
      timestamp:
        description: 错误发生时间
        type: string
    type: object
...
  models.HeartbeatRequest:
    properties:
      config_updated_at:
        description: 客户端配置更新时间，可选
        type: string
      hardware_fingerprint:
        description: 硬件指纹，必填
        type: string
      license_key:
        description: 许可证密钥，必填
        type: string
      software_version:
        description: 软件版本，可选
        type: string
      usage_data:
        additionalProperties: true
        description: 使用数据，可选
        type: object
    required:
    - hardware_fingerprint
    - license_key
    type: object
  models.HeartbeatResponse:
    properties:
      config_updated:
        description: 配置是否有更新
        type: boolean
      heartbeat_interval:
        description: 下次心跳间隔(秒)
        type: integer
      license_file:
        description: base64编码的新许可证文件(如有更新)
        type: string
      status:
        description: 许可证状态
        type: string
    type: object
  models.License:
    properties:
      activated_at:
        type: string
      activation_ip:
        type: string
      authorization_code:
        allOf:
        - $ref: '#/definitions/models.AuthorizationCode'
        description: 关联字段（用于查询时的JOIN）
      authorization_code_id:
        type: string
      config_updated_at:
        type: string
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      customer_id:
        type: string
      device_info:
        type: object
      hardware_fingerprint:
        type: string
      id:
        type: string
      is_online:
        type: boolean
      is_online_display:
        type: string
      last_heartbeat:
        type: string
      last_online_ip:
        type: string
      license_key:
        type: string
      status:
        type: string
      status_display:
        type: string
      updated_at:
        type: string
      usage_data:
        type: object
    type: object
  models.LicenseCreateRequest:
    properties:
      activation_ip:
        description: 激活IP，可选
        type: string
      authorization_code_id:
        description: 授权码ID，必填
        type: string
      device_info:
        additionalProperties: true
        description: 设备信息，可选
        type: object
      hardware_fingerprint:
        description: 硬件指纹，必填
        type: string
    required:
    - authorization_code_id
    - hardware_fingerprint
    type: object
...
  models.LicenseListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/models.LicenseListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  models.LicenseRevokeRequest:
    properties:
      reason:
        description: 撤销原因，可选
        maxLength: 500
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 软件授权管理系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: License Manager API
  version: "1.0"
paths:
...
  /api/v1/activate:
    post:
      consumes:
      - application/json
      description: 客户端使用授权码激活软件，获取许可证文件
      parameters:
      - description: 激活请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 激活成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ActivateResponse'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 授权码不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 授权码已锁定或已过期
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: 激活数量已达上限
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 激活许可证
      tags:
      - 许可证激活
...
  /api/v1/heartbeat:
    post:
      consumes:
      - application/json
      description: 客户端定期发送心跳，更新在线状态和使用数据
      parameters:
      - description: 心跳请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 心跳成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.HeartbeatResponse'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 许可证不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 许可证已被撤销
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 心跳检测
      tags:
      - 许可证激活
  /api/v1/licenses:
    get:
      consumes:
      - application/json
      description: 分页查询许可证列表，支持筛选
      parameters:
      - description: 页码，默认1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页条数，默认20，最大100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 授权码ID筛选
        in: query
        name: authorization_code_id
        type: string
      - description: 客户ID筛选
        in: query
        name: customer_id
        type: string
      - description: 状态筛选
        enum:
        - active
        - inactive
        - revoked
        in: query
        name: status
        type: string
      - description: 在线状态筛选
        in: query
        name: is_online
        type: boolean
      - description: 排序字段，默认created_at
        enum:
        - created_at
        - updated_at
        - activated_at
        - last_heartbeat
        in: query
        name: sort
        type: string
      - description: 排序方向，默认desc
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LicenseListResponse'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 查询许可证列表
      tags:
      - 许可证管理
    post:
      consumes:
      - application/json
      description: 为指定授权码手动创建许可证
      parameters:
      - description: 许可证创建信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LicenseCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.License'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 授权码不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 手动添加许可证
      tags:
      - 许可证管理
  /api/v1/licenses/{id}:
    delete:
      description: 仅允许物理删除状态为revoked的许可证
      parameters:
      - description: 许可证ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 许可证不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 许可证尚未撤销
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除已撤销许可证
      tags:
      - 许可证管理
    get:
      consumes:
      - application/json
      description: 根据许可证ID获取许可证详细信息
      parameters:
      - description: 许可证ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LicenseDetailResponse'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 许可证不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取许可证详情
      tags:
      - 许可证管理
  /api/v1/licenses/{id}/download:
    get:
      consumes:
      - application/json
      description: 下载加密的许可证文件，用于客户端软件激活
      parameters:
      - description: 许可证ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: 包含许可证文件和RSA公钥的压缩包
          schema:
            type: file
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 许可证不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 许可证已被撤销
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 下载许可证文件
      tags:
      - 许可证管理
  /api/v1/licenses/{id}/revoke:
    put:
      consumes:
      - application/json
      description: 撤销指定的许可证，撤销后无法再使用
      parameters:
      - description: 许可证ID
        in: path
        name: id
        required: true
        type: string
      - description: 撤销原因
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LicenseRevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 撤销成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.License'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 许可证不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 许可证已被撤销
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 撤销许可证
      tags:
      - 许可证管理
...
  /api/v1/stats/overview:
    get:
      consumes:
      - application/json
      description: 获取授权码、许可证的总体统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.StatsOverviewResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取授权概览统计
      tags:
      - 统计分析
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"